HumanLLM/requirements.txt
vllm
pandas
pyyaml
packaging
numba
numpy
openai
httpx
scikit-learn
scipy
tokenizers
transformers
tqdm
wandb
multiprocess
azure-cli
azure-identity
tiktoken
accelerate
datasets
trl
peft
liger-kernel
bitsandbytes
deepspeed
sentence_transformers
emoji
langdetect
flash-attn-2.8.1
matplotlib
